+++ src/components/Calendar.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { DragContext } from '../contexts/drag'
+
+const Calendar = ({ calendar }) => (
+    <DragContext>
+        {calendar}
+    </DragContext>
+)
+
+export default Calendar
