+++ src/components/WeeklyCalendar/WeeklyCalendar.js
@@ -6,13 +6,14 @@ import moment from 'moment'
 import ButtonArea from '../ButtonArea/ButtonArea'
 import DragDate from '../Drag/DragDate'
 import { calendarType } from '../../const/drag'
-import { useCalendarContext, useDragDateContext, useDragScheduleContext } from '../../contexts/calendar'
+import { useCalendarContext } from '../../contexts/calendar'
 import useToggle from '../CalendarItem/useToggle'
 import { setCalendar } from '../../reducers/dragDate'
 import { resetScheduleDrag } from '../../reducers/dragSchedule'
 import { updateCalendar } from '../../reducers/calendar'
 import CalendarItemWithPopup from '../CalendarItem/CalendarItemWithPopup'
 import CalendarItemPopupInfo from '../CalendarItem/CalendarItemPopupInfo'
+import { useDragDateContext, useDragScheduleContext } from '../../contexts/drag'
 
 const cx = classNames.bind(styles)
