+++ src/pages/CalendarApp.js
@@ -2,6 +2,7 @@ import React from 'react'
 
 import Monthly from '../components/MonthlyCalendar/Monthly'
 import WeeklyCalendar from '../components/WeeklyCalendar/WeeklyCalendar'
+import Calendar from '../components/Calendar'
 
 const CalendarApp = () => {
 	const monthlyStyle = {
@@ -10,8 +11,8 @@ const CalendarApp = () => {
 
 	return (
 		<>
-			<WeeklyCalendar />
-			<Monthly style={monthlyStyle} />
+			<Calendar calendar={<WeeklyCalendar />} />
+			<Calendar calendar={<Monthly style={monthlyStyle} />} />
 		</>
 	)
 }
